+++ packages/components/src/private-apis.ts
@@ -13,10 +13,13 @@ import Badge from './badge';
 
 import { DateCalendar, DateRangeCalendar, TZDate } from './calendar';
 import {
+	ValidatedCheckboxControl,
+	ValidatedInputControl,
 	ValidatedNumberControl,
 	ValidatedTextControl,
 	ValidatedToggleControl,
 } from './validated-form-controls';
+import { Picker } from './color-picker/picker';
 
 export const privateApis = {};
 lock( privateApis, {
@@ -32,6 +35,9 @@ lock( privateApis, {
 	DateCalendar,
 	DateRangeCalendar,
 	TZDate,
+	Picker,
+	ValidatedInputControl,
+	ValidatedCheckboxControl,
 	ValidatedNumberControl,
 	ValidatedTextControl,
 	ValidatedToggleControl,