+++ packages/components/CHANGELOG.md
@@ -2,11 +2,13 @@
 
 ## Unreleased
 
-## 30.3.1-next.0 (2025-09-05)
-
 ### Bug Fixes
 
 -   `Modal`: Fix modal headings and labels to use proper editor text color instead of wp-admin colors ([#71311](https://github.com/WordPress/gutenberg/pull/71311)).
+
+### Internal
+
+-   `ValidatedCheckboxControl`: Expose the component under private API's ([#71505](https://github.com/WordPress/gutenberg/pull/71505/)).
 
 ## 30.3.0 (2025-09-03)